requirements.txt
# ── Web Framework ────────────────────────────────────────
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.20
starlette>=0.38.0
itsdangerous

# ── Database ─────────────────────────────────────────────
motor>=3.6.0
pymongo>=4.9.0
qdrant-client>=1.11.0
certifi

# ── LLM & AI ────────────────────────────────────────────
google-genai>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
openai-whisper>=20250625

# ── Google APIs ──────────────────────────────────────────
google-api-python-client>=2.146.0
google-auth>=2.49.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0

# ── Document Processing ─────────────────────────────────
pdfplumber>=0.11.0
python-docx>=1.1.0
PyPDF2>=3.0.0
reportlab>=4.0.0

# ── Utilities ────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=24.1.0
APScheduler>=3.10.0
httpx>=0.27.0
torch>=2.0.0
numpy>=2.0.0
passlib[bcrypt]
python-jose[cryptography]
openai

# NOTE: ffmpeg must be installed on the system for audio diarization:
#   macOS: brew install ffmpeg
#   Ubuntu: sudo apt install ffmpeg
